DRTII-791 Move SimulationFormFields class to own file
